requirements.txt
amqp==1.4.9
anyjson==0.3.3
APScheduler==3.7.0
asgiref==3.3.4
automium==0.2.6
automium-web==0.1.1
beautifulsoup4==4.9.3
billiard==3.3.0.23
bs4==0.0.1
celery==3.1.26.post2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
colorful==0.5.4
cryptography==3.4.6
cyaron==0.4.2
data==0.4
decorator==4.4.2
Django==3.1.7
django-apscheduler==0.5.2
django-celery==3.3.1
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-mdeditor==0.1.18
fake-useragent==0.1.11
Flask==1.1.2
funcsigs==1.0.2
h11==0.12.0
html2text==2020.1.16
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==3.0.37
legacy==0.1.6
llvmlite==0.36.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
numba==0.53.1
numpy==1.19.5
pycparser==2.20
Pygments==2.8.1
PyMySQL==1.0.2
pytz==2021.1
requests==2.1.0
scipy==1.5.4
six==1.15.0
soupsieve==2.2.1
sparse==0.4.1
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.4
useragent==0.1.1
uvicorn==0.14.0
Werkzeug==1.0.1